+++ lib/src/widgets/input_field.dart
@@ -4,14 +4,18 @@ import 'package:flutter/material.dart';
 class InputField extends StatefulWidget {
   InputField({
     required this.hintText,
+    required this.width,
     required this.obscureText,
     this.controller,
     required this.bgcolor,
     this.onChanged,
+    this.padding,
     super.key,
   });
 
   Color bgcolor;
+  final EdgeInsets? padding;
+  final double width;
   final String hintText;
   TextEditingController? controller;
   final bool obscureText;
@@ -25,18 +29,18 @@ class _InputFieldState extends State<InputField> {
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: const EdgeInsets.all(6),
-
-      // color: Colors.white,
+      padding: widget.padding ?? const EdgeInsets.all(6),
       decoration: BoxDecoration(
-        border: Border.all(color: const Color.fromARGB(255, 0, 0, 0), width: 2),
+        border: Border.all(
+            color: const Color.fromARGB(255, 0, 0, 0),
+            width: widget.width / 100),
         boxShadow: [
           BoxShadow(
             color: widget.bgcolor,
           ),
         ],
       ),
-      width: 250,
+      width: widget.width,
       child: TextField(
         onChanged: (String value) {
           setState(() {