+++ CHANGELOG.md
@@ -1,4 +1,7 @@
+## 0.1
+
+
 ## 0.0.1
 
-* This is the very first release of Neobrutalist UI Elements library.
-* Shadowy Container, Neo Container and Input Field are added.
+- This is the very first release of Neobrutalist UI Elements library.
+- Shadowy Container, Neo Container and Input Field are added.